cluster timestamp updating
